+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - entity `sensor.[hostname]_unit_system_[unit_name]_active_state`
     for each command-line parameter `--monitor-system-unit [unit_name]`
 - command-line option `--log-level {debug,info,warning,error,critical}`
-- suggest polkit rule when poweroff fails due to
+- suggest polkit rule when poweroff & locking fails due to
   `org.freedesktop.DBus.Error.InteractiveAuthorizationRequired`
   (https://github.com/fphammerle/systemctl-mqtt/issues/67)
 - declare compatibility with `python3.11`, `python3.12` & `python3.13`
@@ -56,6 +56,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - apparmor profile for architectures other than x86_64/amd64
   (`ImportError: Error loading [...]/_gi.cpython-38-aarch64-linux-gnu.so: Permission denied`)
+- fatal `org.freedesktop.DBus.Error.InteractiveAuthorizationRequired` when
+  attempting to lock sessions
 - container image / dockerfile:
   - split `pipenv install` into two stages to speed up image builds
   - `chmod` files copied from host to no longer require `o=rX` perms on host